rename the token module to security
